requirements.txt
streamlit
python-dotenv
gemini
google-generativeai
PyPDF2